interactive dropdown menu, but needs some fixing
