+++ packages/nextjs/app/create/_components/asset_attributes.json
@@ -1,0 +1,112 @@
+[{
+    "assetClass": "art",
+    "attributes": [
+        {
+            "name": "title",
+            "required": true,
+            "inputType": "text",
+            "defaultValue": ""
+        },
+        {
+            "name": "artist",
+            "required": true,
+            "inputType": "text",
+            "defaultValue": ""
+        },
+        {
+            "name": "year",
+            "required": false,
+            "inputType": "number",
+            "defaultValue": 0
+        },
+        {
+            "name": "medium",
+            "required": false,
+            "inputType": "text",
+            "defaultValue": ""
+        },
+        {
+            "name": "dimensions",
+            "required": false,
+            "inputType": "text",
+            "defaultValue": ""
+        }
+    ]
+},
+{
+    "assetClass": "real_estate",
+    "attributes": [
+        {
+            "name": "address",
+            "required": true,
+            "inputType": "text",
+            "defaultValue": ""
+        },
+        {
+            "name": "property_type",
+            "required": true,
+            "inputType": "text",
+            "defaultValue": ""
+        },
+        {
+            "name": "square_footage",
+            "required": false,
+            "inputType": "number",
+            "defaultValue": 0
+        },
+        {
+            "name": "year_built",
+            "required": false,
+            "inputType": "number",
+            "defaultValue": 0
+        },
+        {
+            "name": "bedrooms",
+            "required": false,
+            "inputType": "number",
+            "defaultValue": 0
+        },
+        {
+            "name": "bathrooms",
+            "required": false,
+            "inputType": "number",
+            "defaultValue": 0
+        }
+    ]
+},{
+    "assetClass": "car",
+    "attributes": [
+        {
+            "name": "model",
+            "required": true,
+            "inputType": "text",
+            "defaultValue": ""
+        },
+        {
+            "name": "make",
+            "required": true,
+            "inputType": "text",
+            "defaultValue": ""
+        },
+        {
+            "name": "year",
+            "required": true,
+            "inputType": "number",
+            "defaultValue": 0
+        },
+        {
+            "name": "color",
+            "required": false,
+            "inputType": "text",
+            "defaultValue": ""
+        },
+        {
+            "name": "mileage",
+            "required": false,
+            "inputType": "number",
+            "defaultValue": 0
+        }
+    ]
+}]
+
+
